+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.12.2] - 2025-01-06
+
 ### Fixed
 
 - Remove duplicate slash in file path #67, #79
@@ -354,7 +356,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Intiial release
 
-[unreleased]: https://github.com/tmaegel/ntodotxt/compare/v0.12.1...HEAD
+[unreleased]: https://github.com/tmaegel/ntodotxt/compare/v0.12.2...HEAD
+[0.12.2]: https://github.com/tmaegel/ntodotxt/compare/v0.12.1...v0.12.2
 [0.12.1]: https://github.com/tmaegel/ntodotxt/compare/v0.12.0...v0.12.1
 [0.12.0]: https://github.com/tmaegel/ntodotxt/compare/v0.11.0...v0.12.0
 [0.11.0]: https://github.com/tmaegel/ntodotxt/compare/v0.10.1...v0.11.0